DWITIC_Upload/requirements.txt
streamlit
tensorflow-cpu
pillow
numpy
openai
python-dotenv
